% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\docType{data}
\name{hoxc9data}
\alias{hoxc9data}
\alias{hoxc9alexa}
\alias{hoxc9cy3}
\title{HOXC9 uPBM Data}
\format{PBMExperiment with each column corresponding to an array scan and row corresponding to a spot on a PBM array.}
\source{
\url{https://www.dropbox.com/s/rqdxeu9wcbrv1vs/PBM-HOXC9.zip}
}
\description{
Example universal Protein Binding Microarray (uPBM) data are provided for
wild type HOXC9 human transcription factor (TF) and three allelic variants
(R193K, K195R, and R222W). Arrays were scanned for Alexa488 and Cy3
fluorescence to measure TF binding and baseline double-stranded
DNA (dsDNA) at each probe. Alexa488 scans were taken at multiple laser powers.
The Alexa488 and Cy3 scans are provided as separate PBMExperiment
objects with scans in columns and probes in rows. Scan metadata are included
in the \code{colData} of the objects, and probe information is stored in the
\code{rowData}.
}
\section{Objects}{

  \code{hoxc9alexa}: Alexa488 scans for HOXC9 allelic series


  \code{hoxc9cy3}: Cy3 scans for HOXC9 allelic series
}

\seealso{
hoxc9table
}
\keyword{data}
