% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\docType{data}
\name{hoxc9table}
\alias{hoxc9table}
\title{HOXC9 uPBM Sample Table}
\format{tibble with each row corresponding to single array scan.}
\source{
\url{https://www.dropbox.com/s/rqdxeu9wcbrv1vs/PBM-HOXC9.zip}
}
\description{
Table of universal Protein Binding Microarray (uPBM) scans, both Alexa488 and
Cy3, for an experiment with wild type HOXC9 human transcription factor (TF) and
three allelic variants (R193K, K195R, and R222W). Data for the scans in the table
were loaded using \code{upbm::gpr2PBMExperiment} and are available as data
objects \code{hoxc9alexa} and \code{hoxc9cy3}.
}
\seealso{
hoxc9data
}
\keyword{data}
